refactor(Money): multiply 를 Money 에서 진행.
로직에서 Revenue 전체를 Money 로 묶어서 관리를 진행할까 했으나, Money 가 음수값을 에러 핸들링 부분까지 수정을 해야됨. 돈의 값 자체는 입력 받은 Money 만을 의미하도록

+++ src/main/java/blackjack/domain/Game.java
@@ -2,6 +2,7 @@ package blackjack.domain;
 
 import blackjack.domain.card.Deck;
 import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Money;
 import blackjack.domain.participant.Participant;
 import blackjack.domain.participant.Player;
 
@@ -84,13 +85,13 @@ public class Game {
         }
     }
 
-    public double dealerRevenue() {
+    public int dealerRevenue() {
         return (-1) * totalPlayersRevenue();
     }
 
-    private double totalPlayersRevenue() {
+    private int totalPlayersRevenue() {
         return players.stream()
-                      .mapToDouble(Player::revenue)
+                      .mapToInt(Player::revenue)
                       .sum();
     }
